Update AlertTableViewController, WeatherViewController
Changed some colors on WeatherViewController (removed colors from label backgrounds so it can use the view background to look homogeneous).

Had to replace the Table View Controller for the Alert screen to fix a bug I encountered. Everything remained unchanged. Also added color background for cells based on the type of the alert (used a string checker for that)